Split AST analyzer JS/TS helpers
